+++ src/universe.h
@@ -1,6 +1,8 @@
 #ifndef UNIVERSE_H
 #define UNIVERSE_H
 
+#include <stddef.h>
+
 namespace sf {
 class RenderWindow;
 }
@@ -17,8 +19,8 @@ class Universe
 {
 public:
     virtual ~Universe(){}
-    virtual unsigned int height() = 0;
-    virtual unsigned int width()  = 0;
+    virtual size_t height() = 0;
+    virtual size_t width()  = 0;
     virtual void nextGeneration() = 0;
     virtual void addCell(const Position &pos)   = 0;
     virtual void draw(sf::RenderWindow &window) = 0;
